Switch from block scoping to function scoping.
I prefer function scoping because it lets the programmer work faster and because it helps remove var declarations, which are noise. Cases where nonlocal writes are desired are exceedingly rare; Python didn't even that option until late in its life. I do appreciate being able to lay large amounts of code into a branch of a `match` block in OCaml, but, in that case, it probably improves readability to break that code off as a separate named function. Reporting possible read-before-write vars as errors (to come) stamps out the worst hazard of function scoping.

Remove a few begins that didn't do anything. Salvage one or two tests that remain useful from test_blocks.