Remove implicit blocks (nested lists on indent).
If you want an indent to hold a bunch of expressions, use `begin`. We may add trailing colons on the previous line as sugar to regain our concision.

Implicit nested-list creation meant messiness for representing non-nested lists with whitespace, which may well be desired for data formats or other DSLs.

Something is slightly wrong with 1-list parsing. Add a failing test for that.

Add some musings about functions-as-modules.